[record-minmax] Separate RecordFunction.h
This separates RecordFunction.h to .h and .cpp.

ONE-DCO-1.0-Signed-off-by: Hyukjin Jeong <hj1.jeong@samsung.com>
